move sign-up and sign-in components to the components directory
