+++ serve-soup/src/components/NavBar.js
@@ -6,17 +6,32 @@ class NavBar extends React.Component {
   render() {
     return (
       <div className="nav-container">
-        <Link to="/">
-          <img className="nav-logo" src="./img/logo.png" alt="logo" />
-        </Link>
-        <ul className="nav-bar">
-          <NavLink exact to="/protected" activeClassName="nav-active-link">
+        <div className="FormTitle">
+          <NavLink
+            to="/"
+            activeClassName="FormTitle__Link--Active"
+            className="FormTitle__Link"
+          >
+            I'll Serve-Soup
+          </NavLink>
+        </div>
+        <div className="PageSwitcher">
+          <NavLink
+            exact
+            to="/protected"
+            activeClassName="PageSwitcher__Item--Active"
+            className="PageSwitcher__Item"
+          >
             View Inventory
           </NavLink>
-          <NavLink to="/protected/addItem" activeClassName="nav-active-link">
+          <NavLink
+            to="/protected/addItem"
+            activeClassName="PageSwitcher__Item--Active"
+            className="PageSwitcher__Item"
+          >
             Add item
           </NavLink>
-        </ul>
+        </div>
       </div>
     );
   }
